% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_chill_portion.R
\name{cc_chill_portion}
\alias{cc_chill_portion}
\title{Compute chill portions}
\usage{
cc_chill_portion(temp_ts_numeric, delta_t_h = 1)
}
\arguments{
\item{temp_ts_numeric}{numeric vector of temperatures}

\item{delta_t_h}{number}
}
\description{
Compute chill portions
}
\details{
This computes chill portions
Works only for temperatures every hour or 30 min.
}
