% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_download_yr.R
\name{cc_download_yr}
\alias{cc_download_yr}
\title{Download annual compilations of data from CIMIS}
\usage{
cc_download_yr(
  year,
  station = NULL,
  interval = c("daily", "hourly"),
  units = c("imperial", "metric"),
  data_dir = Sys.getenv("CC_DATADIR"),
  overwrite = FALSE,
  sftp_user = Sys.getenv("CIMIS_SFPT_USR"),
  sftp_pwd = Sys.getenv("CIMIS_SFPT_PWD"),
  unzip = FALSE,
  keep_zip = TRUE
)
}
\arguments{
\item{year}{Year(s), integer}

\item{station}{The station IDs to unzip}

\item{interval}{Frequency of observations (daily or hourly)}

\item{units}{Imperial or metric units}

\item{data_dir}{Data directory}

\item{overwrite}{Overwrite downloaded files}

\item{sftp_user}{A username for the CIMIS SFTP server}

\item{sftp_pwd}{A password for the CIMIS SFTP server}

\item{unzip}{Unzip archives when needed, logical}

\item{keep_zip}{Keep the zip file, logical}
}
\description{
Download annual compilations of data from CIMIS
}
\details{
This will download yearly aggregations of data from the CIMIS SFTP server. As of early 2025,
yearly aggregations are available for 1982-2024. To get a username and password for the CIMIS sftp server, contact CIMIS (see their website).

The files come down as zips (20-25MB) (1982-2022), or annual CSV files (2023-2024). Files will be saved to the data directory.
}
\seealso{
\code{\link[=cc_dwnhly_mth]{cc_dwnhly_mth()}}
}
