% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_dwnhly_mth.R
\name{cc_dwnhly_mth}
\alias{cc_dwnhly_mth}
\title{Download monthly compilations of hourly data from CIMIS}
\usage{
cc_dwnhly_mth(
  month,
  data_dir = Sys.getenv("CC_DATADIR"),
  overwrite = FALSE,
  sftp_user = Sys.getenv("CIMIS_SFPT_USR"),
  sftp_pwd = Sys.getenv("CIMIS_SFPT_PWD"),
  unzip = TRUE,
  stations = NULL,
  keep_zip = TRUE
)
}
\arguments{
\item{month}{months(s), integer}

\item{data_dir}{Data directory}

\item{overwrite}{Overwrite files already downloaded}

\item{sftp_user}{A username for the CIMIS SFTP server}

\item{sftp_pwd}{A password for the CIMIS SFTP server}

\item{unzip}{Unzip archives when needed, logical}

\item{stations}{The station IDs to unzip}

\item{keep_zip}{Keep the zip file, logical}
}
\description{
Download monthly compilations of hourly data from CIMIS for the current calendar year
}
\details{
This will download hourly data for all CIMIS stations in one month chunks.
CIMIS provides monthly compilations for the last 12 months. These monthly archives are
typically uploaded to the server on the first of the month (for the previous month).
To get a username and password for the CIMIS sftp server, contact CIMIS (see their website).
}
\seealso{
\code{\link[=cc_dwnhly_mth]{cc_dwnhly_mth()}}
}
