% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_dwnhly_yr.R
\name{cc_dwnhly_yr}
\alias{cc_dwnhly_yr}
\title{Download annual compilations of hourly data from CIMIS}
\usage{
cc_dwnhly_yr(
  year,
  data_dir = Sys.getenv("CC_DATADIR"),
  overwrite = FALSE,
  sftp_user = Sys.getenv("CIMIS_SFPT_USR"),
  sftp_pwd = Sys.getenv("CIMIS_SFPT_PWD"),
  unzip = FALSE,
  stations = NULL,
  keep_zip = TRUE
)
}
\arguments{
\item{year}{year(s), integer}

\item{data_dir}{Data directory}

\item{overwrite}{Overwrite files already downloaded}

\item{sftp_user}{A username for the CIMIS SFTP server}

\item{sftp_pwd}{A password for the CIMIS SFTP server}

\item{unzip}{Unzip archives when needed, logical}

\item{stations}{The station IDs to unzip}

\item{keep_zip}{Keep the zip file, logical}
}
\description{
Download annual compilations of hourly data from CIMIS
}
\details{
This will download yearly aggregations of hourly data for all CIMIS stations from the CIMIS SFTP server. As of December 2024,
yearly aggregations are available for 1982-2023. To get a username and password for the CIMIS sftp server, contact CIMIS (see their website).

The files come down as zips (20-25MB) (1982-2022), or annual CSV files (2023). Files will be saved to the data directory.
}
\seealso{
\code{\link[=cc_dwnhly_mth]{cc_dwnhly_mth()}}
}
