% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_stack_hly_mth.R
\name{cc_stack_hly_mth}
\alias{cc_stack_hly_mth}
\title{Stack hourly values for a set of stations that have been downloaded by month}
\usage{
cc_stack_hly_mth(
  year,
  month = NULL,
  stations = NULL,
  data_dir = Sys.getenv("CC_DATADIR"),
  proj_subdir = "my_project_area",
  overwrite = FALSE
)
}
\arguments{
\item{year}{Year(s) to stack, integer}

\item{month}{Months(s) to stack, integer}

\item{stations}{The station IDs to stack, integers}

\item{data_dir}{Data directory}

\item{proj_subdir}{Subdirectory of the data directory where the stacked files should go}

\item{overwrite}{Overwrite files already stacked}
}
\description{
Stack hourly values for a set of stations that have been downloaded by month
}
\details{
This will import the hourly values for a set of stations that have been downloaded in monthly aggregations, and append them in
a single CSV file for the year.
This presumes you've already download the monthly aggregations of CIMIS data, and unzipped the CSV files for the stations of interest.
}
