% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_stn_active.R
\name{cc_stn_active}
\alias{cc_stn_active}
\title{Return active CIMIS Stations}
\usage{
cc_stn_active(
  data_dir = Sys.getenv("CC_DATADIR"),
  crs = 3310,
  overwrite = FALSE
)
}
\arguments{
\item{data_dir}{The cache directory}

\item{crs}{The EPSG ID for the returned simple feature point layer}

\item{overwrite}{Overwrite the saved active stations file}
}
\value{
A sf point layer containing the active CIMIS stations
}
\description{
Return active CIMIS Stations
}
\details{
The default value for \code{crs} = 3310 is California Teale Albers.
}
