% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_stn_closest.R
\name{cc_stn_closest}
\alias{cc_stn_closest}
\title{Find the closest CIMIS station to a point}
\usage{
cc_stn_closest(pt, cimis_sf)
}
\arguments{
\item{pt}{either be a sf object or a numeric vector of length 2 containing longitude and latitude coordinates}

\item{cimis_sf}{A sf point object with CIMIS station locations}
}
\description{
Find the closest CIMIS station to a point
}
\seealso{
\link{cc_stn_active}
}
