% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_stn_nn.R
\name{cc_stn_nn}
\alias{cc_stn_nn}
\title{Return a list of CIMIS Station Nearest Neighbors}
\usage{
cc_stn_nn(cimis_sf, nn = 4)
}
\arguments{
\item{cimis_sf}{A simple feature point object of CIMIS stations}

\item{nn}{number of nearest neighbors (integer)}
}
\value{
A list
}
\description{
Return a list of CIMIS Station Nearest Neighbors
}
