% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_stn_wall2wall.R
\name{cc_stn_wall2wall}
\alias{cc_stn_wall2wall}
\title{Identify CIMIS stations for an area}
\usage{
cc_stn_wall2wall(poly, cimis_sf, nn = 1)
}
\arguments{
\item{poly}{A sf polygon object}

\item{cimis_sf}{A sf point object with CIMIS station locations}

\item{nn}{Number of nearest neighbors}
}
\value{
A vector of CIMIS station IDs
}
\description{
Identify CIMIS stations to get wall-to-wall coverage for an area
}
\details{
This function will identify the nearest \code{nn} CIMIS stations for every inch within poly.
}
\seealso{
\link{cc_stn_active}
}
