% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_treetemp.R
\name{cc_treetemp}
\alias{cc_treetemp}
\title{Predict tree temperature}
\usage{
cc_treetemp(
  station,
  nn_list,
  nn = 3,
  start_dt,
  end_dt,
  pt,
  treetempgam_rds,
  data_dir = Sys.getenv("CC_DATADIR"),
  proj_subdir = "my_project_area",
  all_cols = FALSE
)
}
\arguments{
\item{station}{Station id, integer of length 1}

\item{nn_list}{List of nearest neighbors (see \link{cc_stn_nn})}

\item{nn}{Number of nearest neighbors}

\item{start_dt}{Start date}

\item{end_dt}{End date}

\item{pt}{Orchard point}

\item{treetempgam_rds}{Path to the Rds file for the GAM model}

\item{data_dir}{The data directory}

\item{proj_subdir}{Project subdirectory}

\item{all_cols}{Return all columns}
}
\description{
Predict tree temperature
}
\details{
This takes the gap filled data, creates additional variables for the tree temperature gam,
and uses the gam to predict tree temperature (i.e., bark temperature)

This function should be relatively fast (1-2 seconds). It requires that GAP filled data for the
station of interest and the period of interest already be available as a RDS file. See also
cc_fillgaps.

Unlike many of the other functions, cc_treetemp accepts 1 and only 1 station. You must also
pass coordinates of the orchard of interest. Presumably, the station passed is the closest
CIMIS station to

It returns the predicted bark temperature, as well as the chill portions
(in a future version, the chill portions could be made optional).
}
