% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_stn_active.R
\name{cc_stn_active}
\alias{cc_stn_active}
\title{Return active CIMIS Stations}
\usage{
cc_stn_active(
  crs = 3310,
  use_cache = TRUE,
  data_dir = Sys.getenv("CC_DATADIR"),
  overwrite = FALSE
)
}
\arguments{
\item{crs}{The EPSG ID for the returned simple feature point layer}

\item{data_dir}{The cache directory}

\item{overwrite}{Overwrite the saved active stations file}
}
\value{
A sf point layer containing the active CIMIS stations
}
\description{
Return active CIMIS Stations
}
\details{
This will download the active CIMIS stations and return them as a sf object . The default value for \code{crs} = 3310 is California Teale Albers (NAD83).

As a side effect, the function can also save the active stations as a geojson file.
}
